requirements.txt
pymetasploit3
python-nmap
xmltodict
boto3
botocore
